+++ src/Processing/Options.php
@@ -3,8 +3,6 @@
 namespace Orisai\ObjectMapper\Processing;
 
 use Orisai\Exceptions\Logic\InvalidState;
-use Orisai\ObjectMapper\MappedObject;
-use function array_keys;
 use function get_class;
 use function sprintf;
 
@@ -21,9 +19,6 @@ final class Options
 
 	/** @var array<class-string, object> */
 	private array $dynamicContexts = [];
-
-	/** @var array<class-string<MappedObject>, true> */
-	private array $processedClasses = [];
 
 	public function __construct()
 	{
@@ -107,25 +102,6 @@ final class Options
 		}
 
 		return $this->dynamicContexts[$class];
-	}
-
-	/**
-	 * @param class-string<MappedObject> $class
-	 */
-	public function withProcessedClass(string $class): self
-	{
-		$self = clone $this;
-		$self->processedClasses[$class] = true;
-
-		return $self;
-	}
-
-	/**
-	 * @return list<class-string<MappedObject>>
-	 */
-	public function getProcessedClasses(): array
-	{
-		return array_keys($this->processedClasses);
 	}
 
 	/**